LayoutTests/svg/dom/method-argument-aritychecks.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<svg>
  <feDropShadow id=dropshadow></feDropShadow>
  <feGaussianBlur id=gaussian></feGaussianBlur>
  <animate></animate>
  <rect id="foo"></rect>
  <use xlink:href="#foo"></use>
</svg>
<script>
description('Check that incorrect number of arguments throw TypeError.');

var dropShadow = document.getElementById('dropshadow');
var gaussian = document.getElementById('gaussian');
var rect = document.querySelector('rect');
var animateElm = document.querySelector('animate');
var useElm = document.querySelector('use');

debug('SVGAnimationElement');

debug('');
debug('beginElementAt(float offset)');
shouldThrow('animateElm.beginElementAt()');
shouldNotThrow('animateElm.beginElementAt(0)');

debug('');
debug('endElementAt(float offset)');
shouldThrow('animateElm.endElementAt()');
shouldNotThrow('animateElm.endElementAt(0)');

debug('');
debug('');
debug('SVGFEDropShadowElement');

debug('');
debug('setStdDeviation(float stdDeviationX, float stdDeviationY)');
shouldThrow('dropShadow.setStdDeviation()');
shouldThrow('dropShadow.setStdDeviation(0)');
shouldNotThrow('dropShadow.setStdDeviation(0, 0)');

debug('');
debug('');
debug('SVGFEGaussianBlurElement');

debug('');
debug('setStdDeviation(float stdDeviationX, float stdDeviationY)');
shouldThrow('gaussian.setStdDeviation()');
shouldThrow('gaussian.setStdDeviation(0)');
shouldNotThrow('gaussian.setStdDeviation(0, 0)');

</script>
